fix: 실수 타입 수정(float -> double)

+++ group-investment/src/main/java/com/example/group_investment/rule/Rule.java
@@ -1,4 +1,5 @@
 package com.example.group_investment.rule;
+
 import com.example.group_investment.enums.RulePeriod;
 import com.example.group_investment.rule.dto.RuleDto;
 import com.example.group_investment.team.Team;
@@ -19,7 +20,7 @@ public class Rule {
     @JoinColumn(name = "team_id")
     private Team team;
 
-    private float prdyVrssRt;
+    private double prdyVrssRt;
     private int urgentTradeUpvotes;
     private int tradeUpvotes;
     private int depositAmt;
@@ -28,11 +29,11 @@ public class Rule {
     private RulePeriod period;
 
     private Date payDate;
-    private float maxLossRt;
-    private float maxProfitRt;
+    private double maxLossRt;
+    private double maxProfitRt;
 
     @Builder
-    public Rule(Team team, float prdyVrssRt, int urgentTradeUpvotes, int tradeUpvotes, int depositAmt, RulePeriod period, Date payDate, float maxLossRt, float maxProfitRt) {
+    public Rule(Team team, double prdyVrssRt, int urgentTradeUpvotes, int tradeUpvotes, int depositAmt, RulePeriod period, Date payDate, double maxLossRt, double maxProfitRt) {
         this.team = team;
         this.prdyVrssRt = prdyVrssRt;
         this.urgentTradeUpvotes = urgentTradeUpvotes;
